[Refactor] #229 ViewModel 삭제 및 결합
