+++ app/src/main/java/kr/co/lion/unipiece/ui/home/GalleryFragment.kt
@@ -10,10 +10,8 @@ import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import kotlinx.coroutines.launch
-import kr.co.lion.unipiece.R
 import kr.co.lion.unipiece.databinding.FragmentGalleryBinding
-import kr.co.lion.unipiece.ui.MainActivity
-import kr.co.lion.unipiece.ui.home.viewModel.GalleryInfoViewModel
+import kr.co.lion.unipiece.ui.home.viewModel.HomeViewModel
 import kr.co.lion.unipiece.ui.infomation.InfoOneActivity
 import kr.co.lion.unipiece.util.gettingImageName
 import kr.co.lion.unipiece.util.setImage
@@ -22,7 +20,7 @@ class GalleryFragment(val imgRes : String) : Fragment() {
 
     lateinit var fragmentGalleryBinding: FragmentGalleryBinding
 
-    val viewModel: GalleryInfoViewModel by viewModels()
+    val viewModel: HomeViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
