+++ app/src/main/java/kr/co/lion/unipiece/ui/home/HomeFragment.kt
@@ -4,28 +4,21 @@ import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.viewpager2.widget.ViewPager2
-import com.google.rpc.context.AttributeContext.Auth
 import kotlinx.coroutines.launch
 import kr.co.lion.unipiece.R
 import kr.co.lion.unipiece.databinding.FragmentHomeBinding
 import kr.co.lion.unipiece.ui.MainActivity
 import kr.co.lion.unipiece.ui.author.AuthorInfoActivity
-import kr.co.lion.unipiece.ui.author.viewmodel.AuthorInfoViewModel
-import kr.co.lion.unipiece.ui.home.viewModel.GalleryInfoViewModel
 import kr.co.lion.unipiece.ui.home.viewModel.HomeViewModel
-import kr.co.lion.unipiece.ui.home.viewModel.NewsInfoViewModel
-import kr.co.lion.unipiece.ui.home.viewModel.PromoteInfoViewModel
 import kr.co.lion.unipiece.ui.infomation.InfoAllActivity
 import kr.co.lion.unipiece.ui.mypage.VisitGalleryActivity
 import kr.co.lion.unipiece.ui.payment.CartActivity
@@ -40,11 +33,7 @@ class HomeFragment : Fragment() {
 
     lateinit var mainActivity: MainActivity
 
-    val viewModel: PromoteInfoViewModel by viewModels()
-
-    val galleryViewModel:GalleryInfoViewModel by viewModels()
-
-    val homeViewModel: HomeViewModel by viewModels()
+    val viewModel: HomeViewModel by viewModels()
 
 
     val timer = Timer()
@@ -129,7 +118,7 @@ class HomeFragment : Fragment() {
 
             buttonAllGallery.setOnClickListener {
                 lifecycleScope.launch {
-                    val galleryInfo = galleryViewModel.getGalleryDataByDate()
+                    val galleryInfo = viewModel.getGalleryDataByDate()
                     //Log.d("seonguk1234", "${galleryInfo}")
                     val newIntent = Intent(requireActivity(), InfoAllActivity::class.java)
                     newIntent.putExtra("galleryInfo", galleryInfo.toString())
@@ -151,7 +140,6 @@ class HomeFragment : Fragment() {
             viewPagerHomePromote.orientation = ViewPager2.ORIENTATION_HORIZONTAL
             progressBar2.visibility = View.VISIBLE
 
-            val viewModel = ViewModelProvider(this@HomeFragment).get(PromoteInfoViewModel::class.java)
             viewModel.promoteInfo.observe(viewLifecycleOwner, Observer { imageUrls ->
                 bannerVPAdapter.fragmentList.clear()
                 progressBar2.visibility = View.GONE
@@ -175,7 +163,6 @@ class HomeFragment : Fragment() {
             viewPagerHomeNews.adapter = newsVPAdapter
             viewPagerHomeNews.orientation = ViewPager2.ORIENTATION_HORIZONTAL
 
-            val viewModel = ViewModelProvider(this@HomeFragment).get(NewsInfoViewModel::class.java)
             viewModel.newsInfo.observe(viewLifecycleOwner, Observer { imageUrls ->
                 newsVPAdapter.fragmentList.clear()
                 imageUrls.forEach {
@@ -199,7 +186,7 @@ class HomeFragment : Fragment() {
             viewPagerHomeGallery.adapter = galleryVPAdapter
             viewPagerHomeGallery.orientation = ViewPager2.ORIENTATION_HORIZONTAL
 
-            val viewModel = ViewModelProvider(this@HomeFragment).get(GalleryInfoViewModel::class.java)
+
             viewModel.galleryInfoList.observe(viewLifecycleOwner) { imageUrl ->
                 galleryVPAdapter.fragmentList.clear()
                 imageUrl.forEach {
@@ -219,10 +206,10 @@ class HomeFragment : Fragment() {
     private fun initView(){
         fragmentHomeBinding.apply {
             viewLifecycleOwner.lifecycleScope.launch {
-                homeViewModel.authorInfoDataList.observe(viewLifecycleOwner) { value ->
+                viewModel.authorInfoDataList.observe(viewLifecycleOwner) { value ->
                     authorAdapter.updateData(value)
                 }
-                homeViewModel.getAuthorInfoAll()
+                viewModel.getAuthorInfoAll()
             }
         }
     }
