+++ app/src/main/java/kr/co/lion/unipiece/ui/home/NewsFragment.kt
@@ -2,17 +2,15 @@ package kr.co.lion.unipiece.ui.home
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
-import com.bumptech.glide.Glide
 import kotlinx.coroutines.launch
 import kr.co.lion.unipiece.databinding.FragmentNewsBinding
-import kr.co.lion.unipiece.ui.home.viewModel.NewsInfoViewModel
+import kr.co.lion.unipiece.ui.home.viewModel.HomeViewModel
 import kr.co.lion.unipiece.ui.infomation.InfoOneActivity
 import kr.co.lion.unipiece.util.gettingImageName
 import kr.co.lion.unipiece.util.setImage
@@ -21,7 +19,7 @@ class NewsFragment(val imgRes: String) : Fragment() {
 
     lateinit var fragmentNewsBinding: FragmentNewsBinding
 
-    val viewModel: NewsInfoViewModel by viewModels()
+    val viewModel: HomeViewModel by viewModels()
 
 
     override fun onCreateView(
