+++ app/src/main/java/kr/co/lion/unipiece/ui/home/PromoteFragment.kt
@@ -2,27 +2,24 @@ package kr.co.lion.unipiece.ui.home
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
-import com.bumptech.glide.Glide
 import kotlinx.coroutines.launch
 import kr.co.lion.unipiece.databinding.FragmentPromoteBinding
-import kr.co.lion.unipiece.ui.home.viewModel.PromoteInfoViewModel
+import kr.co.lion.unipiece.ui.home.viewModel.HomeViewModel
 import kr.co.lion.unipiece.ui.infomation.InfoOneActivity
 import kr.co.lion.unipiece.util.gettingImageName
 import kr.co.lion.unipiece.util.setImage
-import java.net.URLDecoder
 
 class PromoteFragment(val imgRes: String) : Fragment() {
 
     lateinit var fragmentPromoteBinding: FragmentPromoteBinding
 
-    val viewModel: PromoteInfoViewModel by activityViewModels()
+    val viewModel: HomeViewModel by activityViewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
